+++ test/dag-draw-test-harness-validation-test.el
@@ -14,6 +14,8 @@
 
 ;;; Code:
 
+(add-to-list 'load-path (expand-file-name "test/helpers" (locate-dominating-file default-directory "Eldev")))
+
 (require 'buttercup)
 (require 'dag-draw-test-harness)